requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
contourpy==1.3.0
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
encodec==0.1.1
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
g2pkk==0.1.2
google==3.0.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-dialogflow==2.31.0
googleapis-common-protos==1.65.0
googlesearch-python==1.2.5
grpcio==1.66.1
grpcio-status==1.66.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
huggingface-hub==0.24.6
idna==3.8
inflect==7.4.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mcpi==1.2.1
mdurl==0.1.2
more-itertools==10.5.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
networkx==2.8.8
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
outcome==1.3.0.post0
packaging==24.1
pandas==1.5.3
pillow==10.4.0
pkg==0.2
platformdirs==4.3.2
pooch==1.8.2
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.28.0
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynndescent==0.5.13
pyparsing==3.1.4
pyperclip==1.8.2
pypinyin==0.52.0
PyRect==0.2.0
pysbd==0.3.4
PyScreeze==0.1.30
PySocks==1.7.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.24.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sockets==1.0.0
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
srt==3.5.3
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
trainer==0.0.36
transformers==4.44.2
trio==0.26.2
trio-websocket==0.11.1
triton==3.0.0
TTS==0.22.0
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.2
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
wikipedia==1.4.0
Wikipedia-API==0.7.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.11.1
